Refactor substrait producer into multiple files
